Use explicit integer sizes wherever possible

+++ src/cmd.c
@@ -15,7 +15,7 @@
 #include "spec-device.h"
 
 int ipts_cmd_recv_timeout(struct ipts_context *ipts, enum ipts_command_code code,
-			  struct ipts_response *rsp, int timeout)
+			  struct ipts_response *rsp, u64 timeout)
 {
 	int ret;
